+++ api/schemas/user_affiliations.py
@@ -7,7 +7,7 @@ from app import app
 
 from functions.update_user_role import update_user_role
 
-from model_enums.roles import RoleEnums
+from enums.roles import RoleEnums
 
 from scalars.slug import Slug
 from scalars.email_address import EmailAddress
